+++ DawnLib.Dusk/src/DuskPlugin.cs
@@ -1,12 +1,9 @@
-using System;
 using System.Reflection;
 using BepInEx;
 using BepInEx.Logging;
 using Dawn;
-using Dawn.Internal;
 using Dawn.Utils;
 using Dusk.Internal;
-using Dusk.Utils;
 using UnityEngine;
 
 namespace Dusk;